common ui dev start , deploy public-api
