+++ .gitignore
@@ -38,6 +38,7 @@
 /public-ui/yarn-error.log
 /public-ui/src/generated/*
 /public-ui/src/publicGenerated/*
+/public-ui/sa-key.json
 cloud_sql_proxy
 node_modules
 npm-debug.log
