--- public-ui/public-ui-app.yaml
@@ -1,19 +1,0 @@
-runtime: python27
-api_version: 1
-threadsafe: true
-default_expiration: "2m"
-
-handlers:
-- url: /(.*\.(css|eot|gz|html|ico|jpg|jpeg|js|map|png|svg|ttf|woff|woff2))
-  static_files: dist/\1
-  upload: dist/(.*)
-  secure: always
-- url: /.*
-  static_files: dist/index.html
-  upload: dist/index.html
-  secure: always
-
-# If a file (relative path under public-ui/) matches this regex, do not upload it.
-# Skip everything not starting with "dist".
-skip_files:
-- ^(?!dist).*$
